Backend v2 sec group bug fix + Code restructuring
